Adding a Sallen-Key lowpass filter to the Basic section.
Default settings:
- Freq range approx 100Hz-50kHz
- Butterworth response (12dB/octave)

It has a pair of resistors that are not on textbook SK implementations;
these set an upper bound on the frequency range. Maybe should remove them
and leave it standard?

Limitations:
It would be better with a dual-ganged pot.
I have not been able to figure out whether/how that can be achieved with lcapy.

Squashed from many commits in `slinkp-add-sk-lowpass` branch.